+++ test/w3c/level1/core/hc_nodereplacechildnewchilddiffdocument.js
@@ -143,5 +143,6 @@ function hc_nodereplacechildnewchilddiffdocument() {
 
 
 function runTest() {
+   return; // CSA: disabled; latest DOM spec adopts rather than throws.
    hc_nodereplacechildnewchilddiffdocument();
 }